Resolved conflicts:
- TasksAppPage.tsx: accepted master deletion (replaced by TasksDrawerScreen)
- ConnectionScreen.tsx, GhosttyTerminalLiveKit.tsx, SessionsDrawerScreen.tsx,
  index.tsx: auto-merged cleanly (independent changes on both sides)

+++ packages/tddy-daemon/src/spawner.rs
@@ -354,8 +354,7 @@ pub fn clone_as_user(os_user: &str, git_url: &str, destination: &Path) -> anyhow
                 if libc::setgid(gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
-                #[allow(clippy::cast_possible_wrap)]
-                if libc::initgroups(pw_name.as_ptr(), gid as libc::c_int) != 0 {
+                if libc::initgroups(pw_name.as_ptr(), gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
                 if libc::setuid(uid) != 0 {
@@ -582,8 +581,7 @@ pub fn spawn_as_user(
                 if libc::setgid(gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
-                #[allow(clippy::cast_possible_wrap)]
-                if libc::initgroups(pw_name.as_ptr(), gid as libc::c_int) != 0 {
+                if libc::initgroups(pw_name.as_ptr(), gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
                 if libc::setuid(uid) != 0 {
